pyproject.toml
[project]
name = "platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pulumi-docker>=4.6.2",
    "pulumi-minio>=0.16.3",
    "pulumi-vault>=6.6.0",
    "python-dotenv>=1.1.0",
]
